+++ tests/Controllers/DomainServicesHelperTest.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Bundles\ApiBundle\Tests\Controllers;
+
+use Somnambulist\Bundles\ApiBundle\Tests\Support\Behaviours\BootKernel;
+use Somnambulist\Bundles\ApiBundle\Tests\Support\Stubs\Controllers\TestApiController;
+use Somnambulist\Components\Domain\Commands\CommandBus;
+use Somnambulist\Components\Domain\Jobs\JobQueue;
+use Somnambulist\Components\Domain\Queries\QueryBus;
+use Somnambulist\Components\Domain\Utils\EntityAccessor;
+use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
+
+/**
+ * Class DomainServicesHelperTest
+ *
+ * @package    Somnambulist\Bundles\ApiBundle\Tests\Controllers
+ * @subpackage Somnambulist\Bundles\ApiBundle\Tests\Controllers\DomainServicesHelperTest
+ */
+class DomainServicesHelperTest extends KernelTestCase
+{
+
+    use BootKernel;
+
+    public function testServicesAreRegistered()
+    {
+        /** @var TestApiController $controller */
+        $controller = static::$container->get(TestApiController::class);
+
+        $this->assertTrue(EntityAccessor::call($controller, 'has', $controller, CommandBus::class));
+        $this->assertTrue(EntityAccessor::call($controller, 'has', $controller, JobQueue::class));
+        $this->assertTrue(EntityAccessor::call($controller, 'has', $controller, QueryBus::class));
+    }
+}